Fix support for y4m files with odd dimensions
This fixes the Y4M parser, though yuview eventually refused to read the
file.

Error before: "Error parsing the Y4M header: Could not locate the next
'FRAME' indicator".

Note that the parser cannot deal correctly with the odd-dimensions.

Error after: "With the given settings, the YUV data can not be converted to
RGB: The item width 9 must be divisible by the horizontal subsampling factor
2. The item height 9 must be divisible by the horizontal subsampling factor
2."

Now the problem is that yuview does not seem to support odd dimensions
for subsampled files.